+++ SpriteSheetPacker/SpritePackerProjectFile.h
@@ -13,10 +13,19 @@ class SpritePackerProjectFile
 {
 public:
     SpritePackerProjectFile();
+    virtual ~SpritePackerProjectFile();
 
+    void setAlgorithm(const QString& algorithm) { _algorithm = algorithm; }
+    QString algorithm() const { return _algorithm; }
+
+    void setTrimMode(const QString& trimMode) { _trimMode = trimMode; }
+    QString trimMode() const { return _trimMode; }
 
     void setTrimThreshold(int trimThreshold) { _trimThreshold = trimThreshold; }
     int trimThreshold() const { return _trimThreshold; }
+
+    void setEpsilon(float epsilon) { _epsilon = epsilon; }
+    float epsilon() const { return _epsilon; }
 
     void setTextureBorder(int textureBorder) { _textureBorder = textureBorder; }
     int textureBorder() const { return _textureBorder; }
@@ -56,7 +65,10 @@ public:
     }
 
 protected:
+    QString _algorithm;
+    QString _trimMode;
     int     _trimThreshold;
+    float   _epsilon;
     int     _textureBorder;
     int     _spriteBorder;
     int     _maxTextureSize;