+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 v1.0.3
 ======
+* Add ToolTip for sprite frames
+* Fix bug with copy image (break alpha chanel if image scale)
+* Show identical list of sprite frame
 * Win32: Polished publish dialog (Window title, position)
 * Add display outlines on preview
 * UI and command line implementation of image optimization